+++ codec/protobuf/cast.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/golang/protobuf/proto"
+	"google.golang.org/protobuf/proto"
 )
 
 // cast performs a type assertion on v to verify it is a protocol buffers