+++ codec/protobuf/json.go
@@ -1,74 +1,29 @@
 package protobuf
 
 import (
-	"bytes"
-
-	"github.com/golang/protobuf/jsonpb"
+	"google.golang.org/protobuf/encoding/protojson"
 )
+
+// JSONBasicMediaType is the type and subtype portion of the media-type used to
+// identify data encoded using JSON.
+const JSONBasicMediaType = "application/vnd.google.protobuf+json"
 
 // DefaultJSONMarshaler is the text marshaler used by JSONCodec if none is
 // provided.
-var DefaultJSONMarshaler = &jsonpb.Marshaler{
-	OrigName: true,
+var DefaultJSONMarshaler = protojson.MarshalOptions{
+	UseProtoNames: true,
 }
 
 // DefaultJSONUnmarshaler is the text marshaler used by JSONCodec if none is
 // provided.
-var DefaultJSONUnmarshaler = &jsonpb.Unmarshaler{
-	AllowUnknownFields: true,
+var DefaultJSONUnmarshaler = protojson.UnmarshalOptions{
+	DiscardUnknown: true,
 }
 
-// JSONCodec is an implementation of marshalkit.Codec that marshals protocol
-// buffers messages in text format.
-type JSONCodec struct {
-	commonCodec
-
-	// Marshaler is the JSON marshaler used to marshal messages.
-	// If it is nil, DefaultJSONMarshaler is used.
-	Marshaler *jsonpb.Marshaler
-
-	// Unmarshaler is the JSON unmarshaler used to unmarshal messages.
-	// If it is nil, DefaultJSONUnmarshaler is used.
-	Unmarshaler *jsonpb.Unmarshaler
-}
-
-// BasicMediaType returns the type and subtype portion of the media-type used to
-// identify data encoded by this codec.
-func (c *JSONCodec) BasicMediaType() string {
-	return "application/vnd.google.protobuf+json"
-}
-
-// Marshal returns the binary representation of v.
-func (c *JSONCodec) Marshal(v interface{}) ([]byte, error) {
-	m, err := cast(v)
-	if err != nil {
-		return nil, err
-	}
-
-	jm := c.Marshaler
-	if jm == nil {
-		jm = DefaultJSONMarshaler
-	}
-
-	buf := &bytes.Buffer{}
-	err = jm.Marshal(buf, m)
-	return buf.Bytes(), err
-}
-
-// Unmarshal decodes a binary representation into v.
-func (c *JSONCodec) Unmarshal(data []byte, v interface{}) error {
-	m, err := cast(v)
-	if err != nil {
-		return err
-	}
-
-	ju := c.Unmarshaler
-	if ju == nil {
-		ju = DefaultJSONUnmarshaler
-	}
-
-	return ju.Unmarshal(
-		bytes.NewReader(data),
-		m,
-	)
+// DefaultJSONCodec is a marshalkit.Codec that marshals protocol buffers
+// messages in JSON format.
+var DefaultJSONCodec = Codec{
+	MediaType:   JSONBasicMediaType,
+	Marshaler:   DefaultJSONMarshaler,
+	Unmarshaler: DefaultJSONUnmarshaler,
 }